Refactor settings nickname modal components
