--- apps/frontend/app/components/NicknameSheet/types.ts
@@ -1,4 +1,0 @@
-export interface NicknameSheetProps {
-        initialValue: string;
-        onSave: (value: string) => void;
-}
