+++ apps/frontend/app/components/SettingsListInput/SettingsListInput.tsx
@@ -1,29 +1,26 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React from 'react';
 import { KeyboardAvoidingView, Platform, Text, TextInput, TouchableOpacity, View, Keyboard } from 'react-native';
-import { useTheme } from '@/hooks/useTheme';
-import { useLanguage } from '@/hooks/useLanguage';
 import { useSelector } from 'react-redux';
+
+import { SettingsListInputProps } from './types';
 import styles from './styles';
-import { NicknameSheetProps } from './types';
-import { TranslationKeys } from '@/locales/keys';
+import { useTheme } from '@/hooks/useTheme';
 import { RootState } from '@/redux/reducer';
 import { myContrastColor } from '@/helper/ColorHelper';
 
-const NicknameSheet: React.FC<NicknameSheetProps> = ({ initialValue, onSave }) => {
+const SettingsListInput: React.FC<SettingsListInputProps> = ({
+        placeholder,
+        value,
+        onChangeText,
+        onSave,
+        saveLabel,
+        disableSave = false,
+        autoFocus = true,
+        keyboardType,
+}) => {
         const { theme } = useTheme();
-        const { translate } = useLanguage();
         const { primaryColor, selectedTheme: mode } = useSelector((state: RootState) => state.settings);
         const contrastColor = myContrastColor(primaryColor, theme, mode === 'dark');
-
-        const [value, setValue] = useState(initialValue ?? '');
-
-        useEffect(() => {
-                setValue(initialValue ?? '');
-        }, [initialValue]);
-
-        const trimmedValue = useMemo(() => value?.trim?.() ?? '', [value]);
-
-        const disableSave = useMemo(() => trimmedValue === (initialValue?.trim?.() ?? ''), [initialValue, trimmedValue]);
 
         const Content = (
                 <View
@@ -38,29 +35,30 @@ const NicknameSheet: React.FC<NicknameSheetProps> = ({ initialValue, onSave }) =
                                         backgroundColor: theme.sheet.inputBg,
                                         borderColor: theme.sheet.inputBorder,
                                 }}
-                                autoFocus
-                                placeholder={translate(TranslationKeys.nickname)}
+                                autoFocus={autoFocus}
+                                placeholder={placeholder}
                                 placeholderTextColor={theme.sheet.placeholder}
                                 cursorColor={theme.sheet.text}
                                 selectionColor={primaryColor}
                                 value={value}
-                                onChangeText={setValue}
+                                onChangeText={onChangeText}
+                                keyboardType={keyboardType}
                         />
 
                         <View style={styles.buttonContainer}>
                                 <TouchableOpacity
                                         onPress={() => {
                                                 Keyboard.dismiss();
-                                                onSave(trimmedValue);
+                                                onSave();
                                         }}
-					disabled={disableSave}
+                                        disabled={disableSave}
                                         style={{
                                                 ...styles.saveButton,
                                                 backgroundColor: primaryColor,
                                                 opacity: disableSave ? 0.5 : 1,
                                         }}
                                 >
-                                        <Text style={[styles.buttonText, { color: contrastColor }]}>{translate(TranslationKeys.save)}</Text>
+                                        <Text style={[styles.buttonText, { color: contrastColor }]}>{saveLabel}</Text>
                                 </TouchableOpacity>
                         </View>
                 </View>
@@ -81,4 +79,4 @@ const NicknameSheet: React.FC<NicknameSheetProps> = ({ initialValue, onSave }) =
         );
 };
 
-export default NicknameSheet;
+export default SettingsListInput;
