+++ apps/frontend/app/components/SettingsListInput/index.ts
@@ -1,0 +1,2 @@
+export { default } from './SettingsListInput';
+export type { SettingsListInputProps } from './types';
