+++ apps/frontend/app/components/SettingsListInput/types.ts
@@ -1,0 +1,12 @@
+import { KeyboardTypeOptions } from 'react-native';
+
+export interface SettingsListInputProps {
+        placeholder: string;
+        value: string;
+        onChangeText: (text: string) => void;
+        onSave: () => void;
+        saveLabel: string;
+        disableSave?: boolean;
+        autoFocus?: boolean;
+        keyboardType?: KeyboardTypeOptions;
+}
