+++ apps/frontend/app/components/SettingsListNickname/index.ts
@@ -1,0 +1,2 @@
+export { default } from './SettingsListNickname';
+export type { SettingsListNicknameProps } from './types';
