+++ apps/frontend/app/components/SettingsListNickname/types.ts
@@ -1,0 +1,4 @@
+export interface SettingsListNicknameProps {
+        initialValue?: string;
+        onSave: (value: string) => void;
+}
